Incorporates change requests in PR
